Ajout de commentaires

+++ datamodel/exerciceDAO.js
@@ -14,7 +14,7 @@ module.exports = class ExerciceDAO extends BaseDAO {
 
     getAllFavori(IDUser) {
         return new Promise((resolve, reject) =>
-            this.db.query(`SELECT * FROM Favori_exo WHERE IDUser = '${IDUser}' ORDER BY id`)
+            this.db.query("SELECT * FROM Favori_exo WHERE IDUser = $1 ORDER BY id", [IDUser])
                 .then(res => resolve(res.rows))
                 .catch(e => reject(e)))
     }
@@ -29,34 +29,16 @@ module.exports = class ExerciceDAO extends BaseDAO {
                 .catch(e => reject(e)))
     }
 
-    isValidExo(exercice) {
-        exercice.name = exercice.name.trim()
-        if (exercice.name === "") return false
-        exercice.description = exercice.description.trim()
-        if (exercice.description === "") return false
-        exercice.groupe = exercice.groupe.trim()
-        if (exercice.groupe === "") return false
-        exercice.type = exercice.type.trim()
-        if (exercice.type === "") return false
-        return true
-    }
-
-    update(exercice) {
-        return this.db.query("UPDATE List SET name=$2,description=$3,groupe=$4,type=$5 WHERE id=$1",
-            [exercice.id, exercice.name, exercice.description, exercice.groupe, exercice.type])
-    }
-
     insertExo(exercice){
-        return this.db.query(`INSERT INTO Exercice (name,description,groupe,type) VALUES  ('${exercice.name}','${exercice.description}','${exercice.groupe}','${exercice.type}') RETURNING id`)
+        return this.db.query("INSERT INTO Exercice (name,description,groupe,type) VALUES  ($1,$2,$3,$4) RETURNING id", [exercice.name,exercice.description,exercice.groupe,exercice.type])
     }
 
     insertExoFavori(exercice_id, user_id){
-        return this.db.query(`INSERT INTO Favori_exo (IDUser, IDExo) VALUES  ('${user_id}','${exercice_id}') RETURNING id`)
+        return this.db.query("INSERT INTO Favori_exo (IDUser, IDExo) VALUES  ($1,$2) RETURNING id", [user_id, exercice_id])
     }
 
     deleteExoFavori(exercice_id, user_id){
-        return this.db.query("DELETE FROM Favori_exo WHERE IDExo=$1 AND IDUser=$2",
-            [exercice_id, user_id])
+        return this.db.query("DELETE FROM Favori_exo WHERE IDExo=$1 AND IDUser=$2", [exercice_id, user_id])
     }
 
 }
